perf(bundle): split DOMAIN_MAP out of domains.ts
useFilteredGraph (used by every critical-path component that renders a
filtered-graph view — RiskPropagationFlow, StructuralMetrics, NodeInspector,
all four canvas surfaces) imported DOMAIN_MAP from @/lib/domains, a
333-LOC file. The hook only needs the ~30-LOC lookup table; the rest is
DOMAIN_GROUPS (170 LOC of card metadata + descriptions + icons) and
the DOMAIN_TO_CARD reverse-lookup.

New src/lib/domain-map.ts houses just the DOMAIN_MAP constant.
domains.ts re-imports it and re-exports for backward compat.
useFilteredGraph now imports from the lightweight file directly.

vitest 1524/1524 pass. tsc clean.
